be/models/columns: moves collections from the `transport` package
