+++ backend/models/columns/collections.go
@@ -1,0 +1,6 @@
+package columns
+
+type (
+	GroupIds []GroupId
+	UserIds  []UserId
+)
