+++ backend/models/columns/validate.go
@@ -27,3 +27,19 @@ func (v Phone) Validate() any        { return phoneNumber.Validate(string(v)) }
 func (v SessionToken) Validate() any { return sessionToken.Validate(string(v)) }
 func (v Username) Validate() any     { return username.Validate(string(v)) }
 
+
+func collect[E interface{ Validate() any }](s []E) any {
+	issues := []any{}
+	for _, e := range s {
+		if issue := e.Validate(); issue != nil {
+			issues = append(issues, issue)
+		}
+	}
+	if len(issues) > 0 {
+		return issues
+	}
+	return nil
+}
+
+func (gs GroupIds) Validate() any { return collect(gs) }
+func (us UserIds) Validate() any  { return collect(us) }
